Minor bug fixes to episode 3 (#109)
* cor_mat Heatmap fix

Displaying error "a matrix like object is required as argument to col on website"
Proposed fix: remove cols argument redundancy

* fix syntax in vfold_cv

repetitions -> repeats

* Load ComplexHeatmap package for Heatmaps in episode 3

* remove prep from recipe in tidymodels example

I'm not sure you can use prep in the recipe and set training data and then use the recipe in a workflow. To reproduce, issue evident when running the epsiode code in isolation.