remove all reference to supervised learning
It is referenced in the callout after its first and only mention in the introduction (and not referenced in any other episodes as far as I can tell). I think it's probably not necessary and maybe even distracting from a cognitive overload point of view. If we want to integrate this with some ML jargon, I would suggest it's included much earlier in the episodes and the terminology is used throughout.